Renamed health dashboard file to match naming convention
